feat: Add VideoPlayer reset
