+++ spx-gui/src/pages/guidance/storyline-list.vue
@@ -42,7 +42,7 @@ import CenteredWrapper from '@/components/community/CenteredWrapper.vue'
 import StoryLinesSection from '@/components/guidance/StoryLineSection.vue'
 import StoryLineItem from '@/components/guidance/StoryLineItem.vue'
 import { useQuery } from '@/utils/query'
-import { listStoryLine } from '@/apis/storyline'
+import { listStoryLine } from '@/apis/guidance'
 import { useResponsive } from '@/components/ui'
 
 const isDesktopLarge = useResponsive('desktop-large')
